moved icon text as it was showing through in MagiC
Blank icon text + driver letter was showing through in MagiC.
Moved these to outside of the transparent 'window' area.

Could not find a way to do so with ResourceMaster so restorted to using Interface instead where these can be freely moved within an Icon area.
Hence the removal of RSM file.